UI düzenlemeleri

+++ app/src/main/java/com/yasinsenel/weatherforecast/model/AirQuality.kt
@@ -4,7 +4,7 @@ data class AirQuality(
     val co: Double,
     val gb_defra_index: Int,
     val no2: Double,
-    val o3: Int,
+    val o3: Double,
     val pm10: Double,
     val pm2_5: Double,
     val so2: Double,
